Obsolete timeout storage
