+++ src/NServiceBus.Persistence.AzureStorage/AzureStoragePersistence.cs
@@ -8,7 +8,6 @@ namespace NServiceBus
     {
         AzureStoragePersistence()
         {
-            Supports<StorageType.Timeouts>(s => s.EnableFeatureByDefault<AzureStorageTimeoutPersistence>());
             Supports<StorageType.Sagas>(s => s.EnableFeatureByDefault<AzureStorageSagaPersistence>());
             Supports<StorageType.Subscriptions>(s => s.EnableFeatureByDefault<AzureStorageSubscriptionPersistence>());
         }
