--- src/NServiceBus.Persistence.AzureStorage/Timeout/TimeoutLogic/TimeoutChunkComparer.cs
@@ -1,18 +1,0 @@
-﻿namespace NServiceBus.Timeout.TimeoutLogic
-{
-    using Core;
-    using System.Collections.Generic;
-
-    class TimeoutChunkComparer : IEqualityComparer<TimeoutsChunk.Timeout>
-    {
-        public bool Equals(TimeoutsChunk.Timeout x, TimeoutsChunk.Timeout y)
-        {
-            return x.Id == y.Id && x.DueTime == y.DueTime;
-        }
-
-        public int GetHashCode(TimeoutsChunk.Timeout obj)
-        {
-            return (obj.DueTime.GetHashCode() * 397) ^ (obj.Id?.GetHashCode() ?? 0);
-        }
-    }
-}
